pyproject.toml
[project]
name = "sparklinkai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery[redis]>=5.5.3",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "numpy>=2.3.3",
    "openai>=1.107.2",
    "psutil>=7.0.0",
    "pymilvus>=2.6.1",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "redis>=5.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.35.0",
]
